Made Server a wrapper around NaiveBayes, using double instead of float.
dbus-c++ maps "d" to double.

+++ src/naive-bayes.hh
@@ -16,7 +16,7 @@ namespace wordtip {
         // Extracts the features and multiplies all their probabilities
         // together to get an overall probability.
         // Actually calculates Pr(Document | Category).
-        virtual float get_document_prob(const Glib::ustring& text,
+        virtual double get_document_prob(const Glib::ustring& text,
                                         const Glib::ustring& cat);
 
         // Returns naive Bayesian probability of category cat
@@ -26,19 +26,19 @@ namespace wordtip {
         //
         // Here we ignore Pr(text) as it's the same no matter
         // which item is given.
-        virtual float get_prob(const Glib::ustring& text,
+        virtual double get_prob(const Glib::ustring& text,
                                const Glib::ustring& cat);
 
         // Any resulting classification, when calculated, should be larger
         // then the nearest result by a certain threshold.
-        virtual void set_threshold(const Glib::ustring& cat, float threshold);
-        virtual float get_threshold(const Glib::ustring& cat);
+        virtual void set_threshold(const Glib::ustring& cat, double threshold);
+        virtual double get_threshold(const Glib::ustring& cat);
 
         // The actual Naive Bayes classification method.
         virtual Glib::ustring classify(const Glib::ustring& text, const Glib::ustring& default_cat = Glib::ustring());
 
     protected:
-        std::map<Glib::ustring, float> thresholds_;
+        std::map<Glib::ustring, double> thresholds_;
     };
 
 } // namespace wordtip
